New default cwdevice.
The new default cwdevice is null. It is a default device at startup
(thanks to new function cwdaemon_cwdevice_init()), unless it is
changed through command line, or until it is changed through network
request. It is also a default fallback device if something goes wrong
when changing the cwdevice using network request.