Rewriting code handling PTT delay.
This has started as an attempt to make behaviour of the code the same as in previous version, but turned into more digging, rewriting and improving that I initially intended.
At the end the code is longer, but also better commented and handles different cases.

There is one distinct change: until now when value sent in request was negative, the delay has been set to MAX. Now the value is ignored when handling request, and causes exit() when handling command line arg.