Fixed one issue with switching sound systems.
When a network request for changing sound system is received, and cwdaemon tries to switch to the sound system but fails, it then attempts to fall back to "Null" sound system. Opening a "Null" sound system isn't very complicated, so it should succeed. We may not have a working sound, but at least cwdaemon won't attempt to ask for audio on a closed (un-opened) sound system.
Also some other minor changes in debug messages.