Test code can now start an instance of cwdaemon
Thanks to addition of test/library/process.c, test 001 can be now executed
without manually starting a cwdaemon.

The cwdaemon process is started on default port, so it may collide with
already running instance (if an instance was started independently from
tests). This is something that should be fixed separately.