Using more generic sound system in test code
Using more generic SOUNDCARD in place of PA. This change in repo will save me
some changes in code each time I'm testing the feature on Alpine Linux.
Perhaps I should configure this option through ./configure because on systems
with console buzzer this option won't work, but for now it stays as SOUNDCARD
because it will work for 99% of cases.