Tests: adding function for checking if test env meets some expectations
The main expectation is that the tests are linked against libcw v7 which
doesn't use signals anymore.

In the future there may be also other checks, like checking if a serial
device is present.

It's better to have explicit tests of env before start of a test, than trying
to figure out why a test has failed.