Minor improvements to configure.ac (checking for OS)
Adding piece of code that checks OS, and sets cwdevice based on that.
The piece of code is copied from unixcw.

I still need to verify how it behaves on FreeBSD, and whether I have used
correct device name.

The name of cwdevice is just an initial (default) value that (in future) can
be overridden.