Fixing "deprecated function" issue reported by clang-tidy
The deprecated libcw function used for debugging of libcw was replaced by
newer libcw function. It's very likely that the debugging of libcw from
cwdaemon didn't work in recent releases because the debug threshold was set
to NONE.

I have also added a scenario for manual tests of "--libcwflags" command line
option.