Functional tests: better control of worker threads to detect problems
Sometimes the thread functions may fail, and I need to know about this to
correctly set status of a test. These changes should help me with it a bit.

One such example of failure is when there is no tty cwdevice plugged-in and
cwdevice observer was unable to start - in such case the thread function in
which the observer is running should report an error. Now the error is
reported to main(), and main() can recognize a failed test.