Functional tests: improving test 001 (EXIT request)
Improving handling of exit of child process: adding a handler of SIGCHLD
signal.

Adding a second test case where we send the EXIT request to a cwdaemon that
was just started.

Using recently added events_t to collect and evaluate events that happen
during execution of test case.

Overall changes to the code of the test are quite significant, but the
general idea behind the test is still the same: see how cwdaemon handles EXIT
request.