Functional tests: better logic for controlling duration of receiving
The logic that controlled main loop in receiver's main thread has been
simplified.

It is now simpler, and at the same time it allows us make a decision that no
new characters will be incoming and the loop can terminate.

This decreases time needed to execute a test (we don't have to blindly wait
e.g. 30 seconds for any receive events that may or may not happen).