Functional tests: extending client_send_request()
It's now possible to add 'value' string to escaped requests that don't
require the value.

Adding value for such requests will be necessary in a test that I'm working
on. The test attempts to send misc requests to cwdaemon, some of them
malformed (like an "abort" request that does have some bytes after <ESC>
code).