Tests: reviewing defines used in sending escaped requests
There is a list of defines that are used to specify code of escaped request.
They help avoiding magic number.

This commit is making small adjustments to the defines. Now their values are
the same values of integers/characters that are supposed to be put into
actual request sent to cwdaemon.

The defines are moved to code of cwdaemon - into src/cwdaemon.h. They should
be re-used in daemon that recognizes codes in escaped requests and in client
code that sends those requests.

Slowly, over time, the defines will be used in src/cwdaemon.c.

Also updating places that use these defines.

Also updating documentation and comments with more precise wordings.