clang-tidy: magic numbers
Removing some of magic numbers.

The main reason behind doing it right now is to find and fix all places where
client or receiver buffers have hardcoded small sizes. I need to have
consistently sized buffers in all test code.

And the root cause for this change is that I know that there is a problem in
cwdaemon with processing requests that are close in size to size of
cwdaemon's receive buffer. I need to make sure that my test code will be able
to process data larger than 64 bytes :) Common definitions of buffer sizes
should help me with this.

The magic numbers are caught with clang-tidy.