Functional tests: adding test for testing large 'plain request' messages
I know that cwdaemon has problems with handling large replies (replies to
either <ESC>h requests or to caret requests). Therefore I need to write tests
that test behaviour of cwdaemon with large replies - and with large requests
too, just to have better coverage of the area of large messages.

The test added here sends large plain requests (so just characters to play)
to cwdaemon. The size (count of bytes) in the requests is smaller than, equal
to, or larger than size of cwdaemon's receive buffer.

The sent requests do and don't include terminating NUL, to also test this
aspect of cwdaemon. I need to confirm how cwdaemon handles cases where a
request doesn't include terminating NUL.

The test takes a lot of time to complete, so in the future it may be disabled
in regular runs of "make check", even if functional tests are enabled.