Functional tests: stricter definitions of "expected socket reply" data
I need to be able to exactly specify what reply I'm expecting from cwdaemon.
This includes both contents of the reply and size of the reply.

When declaring the expected reply, I must be able to declare whether it
includes terminating NUL - not all replies may be C strings.

Therefore the type describing a socket reply is no longer a regular C string,
but a combination of array of bytes + count of bytes in the array.

After the type of the reply has been changed, many places in tests code had
to be changed too.