Functional tests: testing server's response to large "esc reply" requests
The test code sends "esc reply" requests that are a bit smaller than, equal
to and larger than cwdaemon's receive and reply buffers.

The tests have been added because I know that in some situations the reply
doesn't include expected \r\n trailer due to a bug in cwdaemon.

The problem has been found thanks to fuzzing tests, and now I'm writing
functionality tests before I get to fixing the source of the problem.

For now the code for the new test is disabled.