Functional tests: refactoring expectations
Code that does evaluation of events was (and still is to some small degree)
repetitive. I have wrapped the common pieces of code that compared expected
data with observed data into helper functions.

The helper functions are now in separate file and are re-used by tests.

This decreases repetitions in the code and makes logs more unified across
tests.