Functional tests: further refactoring of evaluation of events
This time I'm focusing more on the more difficult part: on evaluating count
and type of events, and on looking up the events in events array.

The part is more difficult because it varies between tests.

The commit adds two different ways to achieving goal. The "prettier" way is
used in tests that don't have both Morse and socket receive events. The tests
that do have the two events have one difficulty: the current order of
recording of the two events is badly specified in documentation and probably
is wrong.