Functional tests: adding "basic" tests for plain request
The primary motivation for this addition is to have more than one test
functions in g_tests[] array in main.c.

Before this commit, if I disable request_size_tests() test (because it takes
long to complete), then g_tests[] array becomes empty array, and then a
linter complains.

So I'm adding another test function to the array, this time with basic set of
tests.

Notice that in this commit a lot of code is shared between "basic" and
"request size" tests. This will become a norm for other tests in few future
commits.

You will see that the request_size_tests() function may be disabled in the
next commit.