Functional tests: declaration of expected events is now a part of test case
Each test case now declares an array of expected events. For now only types
of the events are declared - contents of the events is empty (but that may
change in the future).

Today this helps in two ways:

- it's easy to declare how many tests does the test case expect; code
  evaluating the events no longer has to do some strange conditional
  calculations.

- it's easy to declare what exactly types of events are expected and in what
  order.

In the future this may bring the following benefit (when the concept is
expanded further):

- evaluation of events may largely become an iteration over and comparison of
  two arrays. One array with recorded events and one array with expected
  events. The iteration may then compare events' types and events' contents.

  But that's still mostly in future.