Fixing minor memory leak
The memory leak occurred when cwdaemon was switching between cwdevices. A
data structure used to store options of tty device is allocated but was not
de-allocated.

This diff is a partial fix for the problem, fixing major part of the leak.
This part could result in steady increase to allocated (but not deallocated)
memory.

There is still a minor leak where one chunk of memory is not deallocated
throughout entire lifetime of program. Since the chunk is small, and there is
no accumulation of chunks, I postpone the fix till next version of cwdaemon
(till next dev cycle).

Fixing the problem thoroughly will require some deeper investigation into
management of cwdevices. I don't want to do it in this dev cycle.