Release checklist: marking "fuzzy testing" item as done
A simple fuzzy test is now implemented.
The test has uncovered some issues.
The test is included in test scenario for testing a release.
The test can be manually started at any time.
Future directions for extending fuzzy tests are described.

The "fuzzy testing" item planned for this release can be now marked as DONE.