Decreasing usage of options-struct-as-global var
The new options_t structure that holds default value(s) of configuration of
cwdaemon is still a global variable in cwdaemon.c, but it is passed few times
as function argument instead of always being referred to as global variable.

This should help with readability.

For now there is still only one member in that struct, but that will change
in the future.