Retracting one fix to shadowing of 'optarg'
I'm slowly evaluating amount of changes in this release, and I decided that I
want to remove one single fix that introduced a lot of changes to cwdaemon.

The fix was fixing clang-tidy's report about shadowing of 'optarg' variable.

This commit is restoring state of things in this area to state in release
0.12.0.