Bugfix: using <ESC>f request was disabling keying on cwdevice
Using SOUND_SYSTEM Escape request (<ESC>f) resulted in disabling of keying on
keying pin of cwdevice.

This problem was reported to me by a user.

This commit adds a fix.

A functionality test has been written and it confirms that the bug was
present and that this commit is fixing the bug. The test still needs some
polishing, so I will commit it later.