Functional tests: adding manual test for SOUND SYSTEM Escape request
The test is necessary to confirm that recent fix for a problem with switching
a sound system is working as expected.

The test was developed before the fix and the test confirmed that a problem
really exists. When the fix was developed, the test confirmed that the
problem is now gone.

I'm committing the test only now because it needed some more polishing.

This is a "manual" test because supervision of each test case and its results
is necessary: a person must confirm whether a sound is being played using
correct sound system.

The repo still contains a Perl script for tests of the Escape request. In
near future the test will be removed and all features of the Perl test will
be integrated into the new manual test written in C.

The test in C has the advantage of being able to observe cwdevice and to
record events occurring on keying pin (and in the future on ptt pin).