Tests: functional tests now recognize early if cwdevice is present
The functional tests require a valid cwdevice with keying pin to be present
on test machine. If the cwdevice is not present, the test returns with error.

Before this change the lack of cwdevice also caused a failure of a test, but
the failure wasn't immediate: few seconds needed to pass before lack of
cwdevice triggered some error state in a test.
Now the failure of a test is much faster, and "make check" in absence of
cwdevice takes much less time.