Bugfix: handling a case when cwdaemon can't open a sound system
Before the fix the cwdaemon was crashing because of NULL pointer dereference:
a libcw function was receiving a NULL generator pointer. Generator for old
sound system was either not initialized at all (during a start of cwdaemon),
or was de-initialized, and a generator for new sound system was not
initialized either.

Such generator variable was then used by libcw's internal function which led
to crash.

The libcw internal function was fixed today in libcw, but I need to be able
to handle this problem on machines that won't update to this version of libcw
any time soon. This commit fixes this.