Tests: adding test for "PORT Escape request" feature
This Escape request is obsoleted and doesn't do nothing, but as long as it is
present in code I need to test it.

I have created a supervised test for this feature because I simply didn't
want to test it fully manually over and over again in this and future
releases.