Tests: re-doing some tests on FreeBSD
I needed to re-test two items on FreeBSD due to recent changes in code of the
package. The changes were made during tests on Debian Linux.

Unfortunately it turns out that generating gcov reports on FreeBSD is not
that easy to do after all, so I decided to FAIL this particular test and
postpone fixing it till later. There is a ticket for that.