Docs: describing procedure to interrupt tuning (GitHub issue #13)
Partially addressing GitHub issue #13 (about cancelling of tuning procedure).
I'm adding to man page an information that ABORT Escape request can be used
to terminate tuning.

Cancelling/terminating the tuning through ABORT Escape request should be good
enough for now. It is consistent with terminating of keying with the ABORT
Escape request.

Perhaps in the future I will add support for value "0" of TUNE Escape request
as suggested in the ticket. But not in this dev cycle.