Tests: updating tests scenario doc with tests done on Alpine
I won't be doing a full set of tests on Alpine, only a basic sub-set. I think
that doing full set of tests on FreeBSD and on Debian Linux will be enough.

The tests on Alpine are meant to confirm that there are no grave errors that
would prevent someone from compiling and using cwdaemon on Alpine.