Tests: marking remaining tests on Alpine as SKIPPED
I won't be repeating those tests on Alpine. I think that the fact that they
passed on FreeBSD and on Debian Linux is enough.

The tests that were done on Alpine (build system tests, unit tests, fuzzing
tests, unattended functional tests) should be enough to give some basic
confidence in correctness of cwdaemon on Alpine.

Notice that I had to lower tests' receiver speed for unattended tests in
order for the tests to PASS. I believe that this is a sign of a problem with
receiver (or with slow CPU on the test machine), not with cwdaemon itself.